refactor: code review improvements - type safety and error handling
- Remove noexcept from session_id() since mutex lock can throw
- Change MAX_MESSAGES_PER_SECOND to size_t for consistency
- Remove unnecessary session counter overflow check (impossible to trigger)
- Add [[nodiscard]] to send() methods to encourage checking return values
- Add explicit (void) casts for intentionally ignored send() return values
- Add explicit closed state check before continuing to read

+++ src/server/config.hpp
@@ -15,7 +15,7 @@ struct config {
     static inline constexpr size_t MAX_WRITE_QUEUE_SIZE = 100;
     static inline constexpr unsigned short DEFAULT_PORT = 8080;
     static inline constexpr unsigned short DEFAULT_TEST_PORT = 18080;
-    static inline constexpr int MAX_MESSAGES_PER_SECOND = 10;
+    static inline constexpr size_t MAX_MESSAGES_PER_SECOND = 10;
     static inline constexpr int PLACEHOLDER_SEAT = -1;
     static inline constexpr double PLACEHOLDER_STACK = 0.0;
     static inline constexpr size_t MAX_CONNECTIONS = 1000;
